+++ YappyCam.hpp
@@ -17,7 +17,7 @@
 
 #define MIN_FPS 0.1
 #define MAX_FPS 7.0
-#define DEFAULT_FPS 4.0
+#define DEFAULT_FPS 4.5
 
 enum DisplayMode
 {