+++ HISTORY.md
@@ -8,3 +8,4 @@
     - Use H.264/MPEG-4 AVC.
 - 2019.XX.YY ver.0.8
     - Improve message of finalizing.
+    - Skip the frame if not loaded in finalizing.