+++ HISTORY.md
@@ -11,3 +11,4 @@
     - Skip the frame if not loaded in finalizing.
     - Fix handling of file path.
     - Change license to the 3-clause BSD license.
+    - Fullscreen game is supported.