+++ HISTORY.md
@@ -12,3 +12,4 @@
     - Fix handling of file path.
     - Change license to the 3-clause BSD license.
     - Fullscreen game is supported.
+    - "Open folder" feature is added.